+++ changelog.md
@@ -8,7 +8,8 @@
 
 ### Changed
 
-- None
+- [Issue 15](https://github.com/jackdewinter/pyscan/issues/15)
+  - added better control of number of columns on display or quiet mode for no display
 
 ### Fixed